Update borwserlist handling
The issues:
- PR #97, openned from the _no-root-config-file_ branch (by me) was
actually not doing its job completely. Error messages still appear when
there is no root config file.
- The browsers list is cached in the same global variable, in a singleton
(_server.js_), therefore, the last loaded _browserslist_ config file applies
to every file.

The solution:
- Rejections from the _configResolver_ are cought in a newly introduced
_catch()_ branch. These errors are ignore, because they only occur when
there is no config file for a certain path and we do not want to proceed
with the validation of such paths.
- The caching of _browserslist_ configs is handled with an object rather
than a simply array. Each property of the object is a path, and the
value is the respected browser scope.